refactor(dp): move gdoc parsing to Datapage.ts
